+++ lib/app/services/services.dart
@@ -1,4 +1,7 @@
 export 'googlesheets_service.dart';
+export 'roommate_service.dart';
+export 'service_utils.dart';
 export 'storage/hive_storage_service.dart';
 export 'storage/storage_key.dart';
 export 'storage/storage_service.dart';
+export 'task_service.dart';